+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## [1.0.0] - JUNE 2020.
 
 * Initial Flutter SimpleTimer Library Release
+* Added TimerController add and subtract methods